Separate Adaptive engine and GDDV
To better organize code, separate into two classes. One just
implement GDDV parsing and other actually doing adaptive processing.

No functional changes are expected.
