+++ Makefile.am
@@ -83,7 +83,8 @@ thermald_SOURCES = \
 	src/thd_sensor_kbl_g_mcp.cpp \
 	src/thd_zone_kbl_amdgpu.cpp \
 	src/thd_sensor_rapl_power.cpp \
-	src/thd_zone_rapl_power.cpp
+	src/thd_zone_rapl_power.cpp \
+	src/thd_gddv.cpp
 
 man5_MANS = man/thermal-conf.xml.5
 man8_MANS = man/thermald.8
